backend/requirements.txt
alembic
auth-lib-profcomff[fastapi]
fastapi
fastapi-sqlalchemy
gunicorn
logging-profcomff
psycopg2-binary
pydantic[dotenv]
pydantic-settings
SQLAlchemy
uvicorn
pycryptodome
